Fix SonarCloud findings: real bugs, blockers, and new-code smells (#99)
Real bugs:
- JsoniqParser: the 'validate type EQName { }' branch looked ahead for
  "strict" twice, making the type-name form unparseable (unreachable branch)
- SortedNodeSequence: the dedup guard only skipped ONE duplicate, so runs of
  three or more equal nodes leaked duplicates; the surrounding while never
  looped (unconditional return)
- BooleanValue: zero-argument calls fell through to args[0]
  (ArrayIndexOutOfBoundsException) for any name other than true/false
- ExtractFromDateTime: the per-source switches fell through into the next
  source's cast (ClassCastException for unmatched component/source combos)
- IOUtils.getStringFromFile: a single InputStream.read may return fewer
  bytes than the file length, silently NUL-padding the tail — now reads all
- ProfileOperator: shadowed the parent's id with a second counter, so
  profiler node ids were double-incremented and inconsistent

Interrupt/resource handling:
- Re-interrupt on InterruptedException in the dot-tool runners and
  BlockingParallelizer waits; best-effort temp/spill-file delete() results
  are now checked everywhere (S899/S2142)

Confusion-prone shadowing (blockers):
- ForBind.min/max -> splitMin/splitMax; DefaultCtxItem.type -> itemType

Intentional-exactness suppressions:
- new BigDecimal(double/float) in Dbl/Flt/Cast keeps the EXACT binary value
  (BigDecimal.valueOf would substitute the shortest decimal approximation,
  changing xs:decimal/xs:integer cast results) — annotated with rationale
  instead of "fixed"

Smells in recently-changed code: side-effect-free marker consumption in
GDay/GMD/GMon/Tokenizer, QueryException private-constructor reorder (no
dead flag), yearString if/else ladder, flattened Ordering ternary,
StringBuilder.isEmpty, unnamed catch patterns, pattern instanceof,
Grouping/SumAvgAggregator/ConstructedNodeBuilder complexity extractions,
shared "count" constant, HashMap.newHashMap, unused import, redundant
casts, double-brace init replaced by a named local class.

Full suite 1104/0.

+++ src/main/java/io/brackit/query/atomic/YMD.java
@@ -276,7 +276,7 @@ public class YMD extends AbstractDuration {
 
   /** Convert a (negative, years, months) tuple into total signed months. */
   private static long totalMonths(final boolean negative, final short years, final byte months) {
-    final long total = (long) years * 12L + months;
+    final long total = years * 12L + months;
     return negative ? -total : total;
   }
